requirements.txt
torch
torchaudio
openai-whisper
fastapi
uvicorn
python-multipart
requests
pydantic
flask
pywebview 